Change size params to use size_t.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@300166 13f79535-47bb-0310-9956-ffa450edef68

+++ jk/native/common/jk_shm.h
@@ -80,13 +80,13 @@ struct jk_shm_w_rec
     int     recover_wait_time;
     time_t  error_time;
     /* Number of bytes read from remote */
-    int     readed;
+    size_t   readed;
     /* Number of bytes transferred to remote */
-    int     transferred;
+    size_t   transferred;
     /* Number of times the worker was elected */
-    int     elected;
+    size_t   elected;
     /* Number of non 200 responses */
-    int     errors;
+    size_t   errors;
 };
 typedef struct jk_shm_w_rec jk_shm_w_rec_t;
